feat: add featre impact metrics initial area
